api/accounts/views.py
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status
from rest_framework.authtoken.models import Token
from django.contrib.auth import authenticate
from .utils import generate_otp, send_otp_email
from .models import CustomUser


class LoginWithOTP(APIView):
    def post(self, request):
        email = request.data.get("email", "")
        try:
            user = CustomUser.objects.get(email=email)
        except CustomUser.DoesNotExist:
            return Response({"error": "User with this email doesn't exist"}, 
                            status=status.HTTP_404_NOT_FOUND)
        otp = generate_otp()
        user.otp = otp
        user.save()
        
        send_otp_email(email, otp)
        # send_otp_phone(phone_number, otp)
        
        return Response({"message": "OTP has been send to your email."},
                        status=status.HTTP_200_OK
                        )


class ValidateOTP(APIView):
    def post(self, request):
        email = request.data.get("email", "")
        otp = request.data.get("otp", "")
        
        try:
            user = CustomUser.objects.get(email=email)
        except CustomUser.DoesNotExist:
            return Response({'error': 'User with this email does not exist.'}, 
                            status=status.HTTP_404_NOT_FOUND)
        
        if user.otp == otp:
            #  Reset the OTP field after successful validation
            user.otp = None 
            user.save()
            
            # Authenticate the user and create or get an authentication token
            token, _ = Token.objects.get_or_create(user=user)
            
            return Response({"token": token.key}, status=status.HTTP_200_OK)
        else:
            return Response({"error": "Invalid OTP."}, status=status.HTTP_400_BAD_REQUEST)
